feat: atualização do tipo do ID

+++ service/local-layer/data-persistence-service/src/main/java/propagandus/datapersistenceservice/core/repositorys/ReactionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import propagandus.datapersistenceservice.core.models.Reaction;
 
 @Repository
-public interface ReactionRepository extends JpaRepository<Reaction, Long> {
+public interface ReactionRepository extends JpaRepository<Reaction,Integer> {
 }
